[#2189] Fixed tests to work with both mim and pymongo. Ran tests against mongo 1.8

+++ Allura/allura/websetup/bootstrap.py
@@ -199,6 +199,8 @@ def clear_all_database_tables():
     for db in conn.database_names():
         db = conn[db]
         for coll in db.collection_names():
+            if coll == 'system.indexes':
+                continue
             db.drop_collection(coll)
